Use pathlib and handle tmp_dir for tests automatically
