+++ tests/trestle/core/models/element_test.py
@@ -21,76 +21,6 @@ from trestle.core.err import TrestleError
 from trestle.core.models.elements import Element, ElementPath
 from trestle.oscal import target
 
-
-def test_element_path_init(sample_target: target.TargetDefinition):
-    """Test element path construction."""
-    assert ElementPath('metadata.title').get() == ['metadata', 'title']
-    assert ElementPath('targets.*').get() == ['targets', '*']
-    assert ElementPath('targets.0').get() == ['targets', '0']
-    assert ElementPath('metadata.parties.0.uuid').get() == ['metadata', 'parties', '0', 'uuid']
-
-    # expect error
-    try:
-        ElementPath('*')
-    except TrestleError:
-        pass
-
-    # expect error
-    try:
-        ElementPath('*.*')
-    except TrestleError:
-        pass
-
-    # expect error
-    try:
-        ElementPath('.')
-    except TrestleError:
-        pass
-
-    # expect error
-    try:
-        ElementPath('.*')
-    except TrestleError:
-        pass
-
-    # expect error
-    try:
-        ElementPath('catalog.groups.*.controls.*')
-    except TrestleError:
-        pass
-
-    # expect error
-    try:
-        ElementPath('metadata..title')
-    except TrestleError:
-        pass
-
-
-def test_element_path_get_element_name(sample_target: target.TargetDefinition):
-    """Test get element name method."""
-    assert ElementPath('metadata.title').get_element_name() == 'title'
-    assert ElementPath('metadata').get_element_name() == 'metadata'
-    assert ElementPath('metadata.parties.*').get_element_name() == 'parties'
-
-
-def test_element_path_get_parent(sample_target: target.TargetDefinition):
-    """Test get parent path method."""
-    assert ElementPath('metadata.title').get_parent_path() == ElementPath('metadata')
-    assert ElementPath('metadata').get_parent_path() is None
-    assert ElementPath('metadata.parties.*').get_parent_path() == ElementPath('metadata.parties')
-    assert ElementPath('metadata.*').get_parent_path() == ElementPath('metadata')
-
-
-def test_element_path_get(sample_target: target.TargetDefinition):
-    """Test get method of element path."""
-    assert ElementPath('metadata').get() == ['metadata']
-    assert ElementPath('metadata.title').get() == ['metadata', 'title']
-
-    assert ElementPath('metadata.title').get_first() == 'metadata'
-    assert ElementPath('metadata.title').get_last() == 'title'
-    assert ElementPath('metadata').get_last() == 'metadata'
-
-
 def test_element_get_at(sample_target: target.TargetDefinition):
     """Test element get method."""
     element = Element(sample_target)
@@ -167,16 +97,3 @@ def test_element_str(sample_target):
     """Test for magic method str."""
     element = Element(sample_target)
     assert str(element) == 'TargetDefinition'
-
-
-def test_element_path_str():
-    """Test for magic method str."""
-    element_path = ElementPath('target.metadata')
-    assert str(element_path) == 'target.metadata'
-
-
-def test_element_path_eq(sample_target):
-    """Test for magic method eq."""
-    assert ElementPath('target.metadata') == ElementPath('target.metadata')
-    assert not(ElementPath('target.metadata') == ElementPath('target.title'))
-    assert not(ElementPath('target.metadata') == Element(sample_target))
