+++ tests/trestle/core/models/plans_test.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Tests for trestle plans module."""
-import os
+import pathlib
 
 from trestle.core.models.actions import FileContentType, WriteFileAction
 from trestle.core.models.elements import Element
@@ -26,22 +26,24 @@ def test_plan_execution(tmp_dir, sample_target: target.TargetDefinition):
     """Test successful execution of a valid plan."""
     content_type = FileContentType.YAML
 
-    base_dir = os.path.join(tmp_dir, 'mytarget')
-    targets_dir = os.path.join(base_dir, 'targets')
+    base_dir = pathlib.Path.joinpath(tmp_dir, 'mytarget')
+    targets_dir = pathlib.Path.joinpath(base_dir, 'targets')
 
     # hand craft a split plan
     split_plan = Plan()
     split_plan.add_action(
-        WriteFileAction(os.path.join(base_dir, 'metadata.json'), Element(sample_target.metadata), content_type)
+        WriteFileAction(pathlib.Path.joinpath(base_dir, 'metadata.json'), Element(sample_target.metadata), content_type)
     )
 
     for tid, t in sample_target.targets.items():
-        split_plan.add_action(WriteFileAction(os.path.join(targets_dir, tid), Element(t), content_type))
+        split_plan.add_action(WriteFileAction(pathlib.Path.joinpath(targets_dir, tid), Element(t), content_type))
 
     # execute the plan
     split_plan.execute()
     split_plan.rollback()
-    os.removedirs(targets_dir)
+
+    pathlib.Path.rmdir(targets_dir)
+    pathlib.Path.rmdir(base_dir)
 
 
 def test_plan_execution_failure():
