+++ tests/trestle/parsing/parsing_test.py
@@ -16,7 +16,7 @@
 """Tests for json and yaml manipulations of oscal files."""
 
 import datetime
-import os
+import pathlib
 
 from dateutil.parser import isoparse
 
@@ -29,8 +29,8 @@ from trestle.utils import fs
 
 import yaml
 
-yaml_path = 'tests/data/yaml/'
-json_path = 'tests/data/json/'
+yaml_path = pathlib.Path('tests/data/yaml/')
+json_path = pathlib.Path('tests/data/json/')
 encoding = 'utf8'
 loader = yaml.Loader
 
@@ -48,13 +48,13 @@ def times_equal(t1, t2):
 def test_yaml_load():
     """Test yaml load."""
     # happy path
-    with open(yaml_path + 'good_simple.yaml', 'r', encoding=encoding) as read_file:
+    with open(pathlib.Path.joinpath(yaml_path, 'good_simple.yaml'), 'r', encoding=encoding) as read_file:
         obj = yaml.load(read_file, Loader=loader)
     assert obj is not None
 
     # unhappy path
     with pytest.raises(yaml.parser.ParserError):
-        with open(yaml_path + 'bad_simple.yaml', 'r', encoding=encoding) as read_file:
+        with open(pathlib.Path.joinpath(yaml_path, 'bad_simple.yaml'), 'r', encoding=encoding) as read_file:
             obj = yaml.load(read_file, Loader=loader)
 
 
@@ -63,13 +63,13 @@ def test_yaml_dump(tmp_dir):
     target_name = 'good_target.yaml'
 
     # happy path
-    with open(yaml_path + target_name, 'r', encoding=encoding) as read_file:
+    with open(pathlib.Path.joinpath(yaml_path, target_name), 'r', encoding=encoding) as read_file:
         target = yaml.load(read_file, Loader=loader)
     assert target is not None
 
     fs.ensure_directory(tmp_dir)
 
-    dump_name = tmp_dir + '/' + target_name
+    dump_name = pathlib.Path.joinpath(tmp_dir, target_name)
 
     with open(dump_name, 'w', encoding=encoding) as write_file:
         yaml.dump(target, write_file)
@@ -81,8 +81,8 @@ def test_yaml_dump(tmp_dir):
     assert saved_target == target
 
     # clean up
-    os.remove(dump_name)
-    os.removedirs(tmp_dir)
+    pathlib.Path.unlink(dump_name)
+    pathlib.Path.rmdir(tmp_dir)
 
 
 def test_oscal_model(tmp_dir):
@@ -90,7 +90,7 @@ def test_oscal_model(tmp_dir):
     good_target_name = 'good_target.yaml'
 
     # load good target
-    with open(yaml_path + good_target_name, 'r', encoding=encoding) as read_file:
+    with open(pathlib.Path.joinpath(yaml_path, good_target_name), 'r', encoding=encoding) as read_file:
         yaml_target = yaml.load(read_file, Loader=loader)
     assert yaml_target is not None
 
@@ -102,7 +102,7 @@ def test_oscal_model(tmp_dir):
     yaml_target_def = yaml_target['target-definition']
 
     fs.ensure_directory(tmp_dir)
-    dump_name = tmp_dir + '/' + good_target_name
+    dump_name = pathlib.Path.joinpath(tmp_dir, good_target_name)
 
     # write the oscal target def out as yaml
     # exclude all None's and rename to original OSCAL form with - instead of _
@@ -114,8 +114,8 @@ def test_oscal_model(tmp_dir):
         oscal_target_def_reload = yaml.load(read_file, Loader=loader)
 
     # clean up
-    os.remove(dump_name)
-    os.removedirs(tmp_dir)
+    pathlib.Path.unlink(dump_name)
+    pathlib.Path.rmdir(tmp_dir)
 
     assert oscal_target_def_reload is not None
 
@@ -129,7 +129,7 @@ def test_oscal_model(tmp_dir):
         assert (times_equal(d[2][0], d[2][1]))
 
     # load good target with different timezone
-    with open(yaml_path + 'good_target_diff_tz.yaml', 'r', encoding=encoding) as read_file:
+    with open(pathlib.Path.joinpath(yaml_path, 'good_target_diff_tz.yaml'), 'r', encoding=encoding) as read_file:
         yaml_target_diff_tz = yaml.load(read_file, Loader=loader)
     assert yaml_target_diff_tz is not None
